Update to fix use of `load_data` function.
Note that previously when plot_results.py does `from process_data import load_data`, it RUNS the whole script. The `if __name__ == '__main__'` line fixes this, this is a python convention that is widely used.

Also, the `load_data` function previously would have worked if you ran the script first (as above), but it might have gotten confused if the .h5 files got into different states than expected. Thus, I've changed it to load the .VEC in the for-loop, then use `load_data` only as an external function.